refactor: move api utility isNameUnique to core as isResourceNameUnique
