+++ packages/amplify-cli-core/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './fileSize';
+export * from './isResourceNameUnique';
 export * from './open';
 export * from './packageManager';
 export * from './recursiveOmit';
